refactor(core): split bind-layer validation and panel layout paths
Extract field helpers and geom/stat contracts from bind-layer, and split
facet vs single-panel placement out of computePanelLayout. Characterization
tests pin rule forms, color-on-fill warnings, free_y axes, and coord-flip labs.
